+++ src/components/buttons/TipsButton.vue
@@ -5,19 +5,20 @@
     :effect="tipsType"
     :placement="placement"
   >
-    <el-button v-bind="$attrs" @click="onClick"><slot /></el-button>
+    <el-button v-bind="$attrs" @click="onClick"><slot>{{ text }}</slot></el-button>
   </el-tooltip>
-  <el-button v-else v-bind="$attrs" @click="onClick"><slot /></el-button>
+  <el-button v-else v-bind="$attrs" @click="onClick"><slot>{{ text }}</slot></el-button>
 </template>
 
 <script setup lang="ts">
 import type { PropType } from 'vue'
-import { filterProp } from '../common/props'
+import { filterProp, buttonTextProp } from '../common/props'
 
 defineOptions({ name: 'WdTipsButton', inheritAttrs: false })
 
 const props = defineProps({
   ...filterProp,
+  ...buttonTextProp,
   /** 悬停提示内容 */
   tips: { type: String, default: '' },
   /** tooltip 主题：dark/light */